renaming NotificationIcon to match MessageIcon
